moved all functionalities to utils.
