+++ aiInferenceModel/ai_models/nuclickSegmentation.py
@@ -18,18 +18,18 @@ def run_ai_model_inferencing(json_data, network):
     image_data = json_data.get("image")
     foreground_data = json_data.get("nuclei_location")
     size_data = json_data.get("tilesize")
-    gx,gy,_,_,x,y = size_data
+    gx, gy, _, _, x, y = size_data
     image = np.array(image_data)
     temp_dir = tempfile.mkdtemp()
 
     # infereing nuclei location #TODO
-    print('Tile reference',size_data)
+    print('Tile reference', size_data)
     print("\n")
     print('nuclei locations', foreground_data)
     print("\n")
     print("image size", image.shape)
 
-    #adding tile reference to the input cordinates
+    # adding tile reference to the input cordinates
     for element in foreground_data:
         element[0] = int(np.abs(element[0] - x))
         element[1] = int(np.abs(element[1] - y))
@@ -97,8 +97,25 @@ def run_ai_model_inferencing(json_data, network):
             # generate contours
             zero_image = np.zeros(output_predictions["pred"].shape)
 
-            zero_image[int(cy-width/2):int(cy+width/2), int(cx-height/2):int(cx + height/2)] = output_predictions["pred"][int(cy-width/2):int(cy+width/2), int(cx-height/2):int(cx + height/2)]
-            contours, _ = cv2.findContours(zero_image.astype('uint8'), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+            zero_image[int(cy -
+                           width /
+                           2):int(cy +
+                                  width /
+                                  2), int(cx -
+                                          height /
+                                          2):int(cx +
+                                                 height /
+                                                 2)] = output_predictions["pred"][int(cy -
+                                                                                      width /
+                                                                                      2):int(cy +
+                                                                                             width /
+                                                                                             2), int(cx -
+                                                                                                     height /
+                                                                                                     2):int(cx +
+                                                                                                            height /
+                                                                                                            2)]
+            contours, _ = cv2.findContours(zero_image.astype(
+                'uint8'), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
 
             output_list = [[[x[0][1] + gx, x[0][0] + gy, 0]
                             for x in arr.tolist()] for arr in list(contours)]
